Add line offset configuration on parse error
▶ This config is useful for stream parsing
▶ Use on subcommand `eval`
▶ Use on subcommand `code` with `--stream` flag

+++ packages/core/src/core/builder.rs
@@ -5,7 +5,7 @@ use pest_derive::Parser;
 #[grammar = "grammar.pest"]
 pub struct Scdlang<'g> {
 	pub(super) path: Option<&'g str>,
-	pub(super) line: Option<u16>,
+	pub(super) line: Option<usize>,
 }
 
 impl<'g> Builder<'g> for Scdlang<'g> {
@@ -13,7 +13,7 @@ impl<'g> Builder<'g> for Scdlang<'g> {
 		self.path = Some(path);
 	}
 
-	fn with_err_line(&mut self, line: u16) {
+	fn with_err_line(&mut self, line: usize) {
 		self.line = Some(line);
 	}
 }
